+++ CHANGELOG.md
@@ -6,7 +6,7 @@
   * Update PPRiskMagnes to 5.6.0 
   * Add `collectDeviceDataOnSuccess(riskCorrelationID: completion:)` method that ensures Magnes SDK completes data submission prior to returning device data.
 * BraintreePayPal
-    * Pass `is_billing_agreement`, `is_purchase`, and `billing_plan_type` to analytics events.
+    * Pass `is_billing_agreement`, `is_purchase`, and `billing_plan_type` to analytics events in order to track the billing with purchase flows.
 
 ## 7.3.0 (2026-01-29)