+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # Braintree iOS SDK Release Notes
 
-## unreleased
-
+## 7.4.0 (2026-02-12)
+* BraintreeVenmo
+    * Expose the `riskCorrelationID` property on `BTVenmoRequest` 
 * BraintreeDataCollector
   * Update PPRiskMagnes to 5.6.0 
   * Add `collectDeviceDataOnSuccess(riskCorrelationID: completion:)` method that ensures Magnes SDK completes data submission prior to returning device data.
@@ -9,7 +10,6 @@
     * Pass `is_billing_agreement`, `is_purchase`, and `billing_plan_type` to analytics events in order to track the billing with purchase flows.
 
 ## 7.3.0 (2026-01-29)
-
 * BraintreePayPal
     * Pass `fundingSource` to the app switch url link and to analytics events. 
     * Fix bug where setting an empty `BTPayPalVaultRequest.userAuthenticationEmail` resulted in an error